refactor: mv reconcile method to check controller
